fix: prevent login error when using module slot
If a user has only Public (module protected) objects they can
choose the module slot as their token. The new v1.0 API now tries
to login regardless if the user provided a pin. This will return
an error when using the module slot as it does not require a login.

I don't know the full intent of the original change to the API but
I figured it would be better to continue to call Login and just ignore
the error to not break current behavior.

+++ crypto11.go
@@ -310,7 +310,9 @@ func Configure(config *Config) (*Context, error) {
 		return nil, errors.WithMessagef(err, "failed to create long term session")
 	}
 	err = instance.ctx.Login(instance.persistentSession, pkcs11.CKU_USER, instance.cfg.Pin)
-	if err != nil {
+
+	// the login will fail when using a module slot so only error out when the token requires a login
+	if err != nil && (instance.token.Flags&pkcs11.CKF_LOGIN_REQUIRED) == pkcs11.CKF_LOGIN_REQUIRED {
 		_ = instance.ctx.Finalize()
 		instance.ctx.Destroy()
 		return nil, errors.WithMessagef(err, "failed to log into long term session")
